requirements.txt
jsonschema>=4.21.1
PyYAML>=6.0.1
fastapi>=0.112.0
uvicorn>=0.30.0
prometheus_client>=0.20.0
pydantic>=2.8.0
chromadb>=0.4.22
sentence-transformers>=2.3.0
tiktoken>=0.5.2
openai>=1.0.0
anthropic>=0.18.0
google-generativeai
python-dotenv>=1.0.0
python-multipart>=0.0.9
